add 12.5 and update 12.2 12.3

+++ Lab12.3/ArrrayProcessing.h
@@ -5,6 +5,8 @@
 #ifndef LAB12_3_ARRRAYPROCESSING_H
 #define LAB12_3_ARRRAYPROCESSING_H
 
-int DeleteEvenNumbers(int array[], int size);
+#include <stddef.h>
+
+size_t DeleteEvenNumbers(int array[], size_t size);
 
 #endif //LAB12_3_ARRRAYPROCESSING_H
