+++ Lab12.3/main.c
@@ -6,11 +6,11 @@ int main() {
     printf("12.3)n=");
     scanf("%i", &n);
     int a[n];
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < n; ++i) {
         scanf("%i", &a[i]);
     }
-    n = DeleteEvenNumbers(a, n);
-    for (int i = 0; i < n; ++i) {
+    n = (int)(DeleteEvenNumbers(a, n));
+    for (size_t i = 0; i < n; ++i) {
         printf("%i ", a[i]);
     }
 }
